+++ src/examples/mvs-kinase-story/context.ts
