+++ src/examples/mvs-kinase-story/elements/snapshot-markdown.tsx
