+++ src/examples/mvs-kinase-story/elements/viewer.tsx
