+++ src/examples/mvs-kinase-story/index.html
@@ -33,7 +33,6 @@
         }
     </style>
     <link rel="stylesheet" type="text/css" href="molstar.css" />
-    <!-- link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/skeleton/2.0.4/skeleton.min.css" integrity="sha512-EZLkOqwILORob+p0BXZc+Vm3RgJBOe1Iq/0fiI7r/wJgzOFZMlsqTa29UEl6v6U6gsV4uIpsNZoV32YZqrCRCQ==" crossorigin="anonymous" referrerpolicy="no-referrer" /-->
     <script type="text/javascript" src="./index.js"></script>
 </head>
