+++ src/examples/mvs-kinase-story/styles.scss
